modules/api/requirements.txt
fastapi
uvicorn
sqlmodel
asyncpg
greenlet
pydantic-settings
httpx
pytest
python-jose[cryptography]
python-dotenv